Updated hokulea (fix for holesky support) (#30)
Bump hokulea to version with Holesky support

Also bumps celo-kona to have a matching hokulea version

Some types were renamed in the hokulea update so some changes were
required to account for that here.

Note also that the implementation of CeloRollupConfig in celo-kona has
been simplified which resulted in some additional changes.